chore #121 - Extract out media selector utils
Closes #121
