+++ package.json
@@ -8,7 +8,7 @@
 		"url": "ssh://git@github.com:dassana-io/web-utils",
 		"directory": "@dassana-io/web-utils"
 	},
-	"version": "0.8.12",
+	"version": "0.8.13",
 	"dependencies": {
 		"axios": "^0.21.1",
 		"axios-retry": "^3.1.9",
